requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.53.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
autogen-agentchat==0.6.1
autogen-core==0.6.1
autogen-ext==0.6.1
av==14.4.0
azure-ai-agents==1.0.1
azure-ai-projects==1.0.0b11
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
bs4==0.0.2
build==1.3.0
cachetools==6.2.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.3.0
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.3
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.34.0
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
ecdsa==0.19.1
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.2
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.41.1
google-crc32c==1.7.1
googleapis-common-protos==1.71.0
gradio==5.33.1
gradio-client==1.10.3
greenlet==3.2.3
griffe==1.7.3
groovy==0.1.2
grpcio==1.70.0
h11==0.16.0
hf-xet==1.1.3
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib-metadata==8.7.0
importlib-resources==6.5.2
ipinfo==5.2.1
ipython==9.3.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter-core==5.9.1
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==1.0.3
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-huggingface==1.0.0
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
lazy-object-proxy==1.11.0
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mcp==1.11.0
mcp-server-fetch==2025.1.17
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
mypy-extensions==1.1.0
narwhals==1.42.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.0
nvidia-cublas-cu12==12.8.4.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-cupti-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cuda-runtime-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cudnn-cu12==9.10.2.21 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cufft-cu12==11.3.3.83 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cufile-cu12==1.13.1.3 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-curand-cu12==10.3.9.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusolver-cu12==11.7.3.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusparse-cu12==12.5.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-cusparselt-cu12==0.7.1 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nccl-cu12==2.27.5 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nvjitlink-cu12==12.8.93 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nvshmem-cu12==3.3.20 ; platform_machine == 'x86_64' and sys_platform == 'linux'
nvidia-nvtx-cu12==12.8.90 ; platform_machine == 'x86_64' and sys_platform == 'linux'
oauthlib==3.3.1
ollama==0.6.0
onnxruntime==1.23.2
openai==1.107.1
openai-agents==0.3.3
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pexpect==4.9.0
pillow==11.2.1
pip==25.3
pip-system-certs==5.3
platformdirs==4.5.0
playwright==1.52.0
plotly==6.1.2
polygon-api-client==1.14.6
posthog==5.4.0
prance==25.4.8.0
prompt-toolkit==3.0.51
propcache==0.3.2
protego==0.4.0
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybars4==0.9.13
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.11.0
pydub==0.25.1
pyee==13.0.0
pygments==2.19.1
pyjwt==2.10.1
pylibsrtp==0.12.0
pymeta3==0.5.1
pyopenssl==25.1.0
pyparsing==3.2.5
pypdf==5.6.0
pypdf2==3.0.1
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4 ; sys_platform == 'win32'
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
pywin32==311 ; sys_platform == 'win32'
pyyaml==6.0.2
qrcode==8.2
readabilipy==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
ruff==0.11.13
redis==6.4.0
redisvl==0.8.2
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-kernel==1.32.2
semantic-version==2.10.0
sendgrid==6.12.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smithery==0.1.0
sniffio==1.3.1
soupsieve==2.7
speedtest-cli==2.1.3
sqlalchemy==2.0.41
sqlite-vec==0.1.6
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
strenum==0.4.15
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
triton==3.5.0 ; platform_machine == 'x86_64' and sys_platform == 'linux'
typer==0.16.0
types-requests==2.32.0.20250602
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tavily==1.1.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.9.0
websockets==14.2
werkzeug==3.1.1
widgetsnbextension==4.0.14
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0